+++ espnet/opts/chainer_backend/adagrad.py
@@ -2,7 +2,7 @@ from argparse import Namespace
 
 from chainer.optimizers import AdaGrad as AdaGrad_chainer
 
-from espnet.optimizers.chainer_backend.opt_interface import OptInterface
+from espnet.opts.chainer_backend.opt_interface import OptInterface
 
 
 class AdaGrad(OptInterface):
